+++ basic_reply.py
@@ -24,6 +24,17 @@ def get_button_of_help(msguser: MessageUser):
     return pld
 
 
+def get_button_of_source_code(msguser: MessageUser):
+    mixin_number = None if not msguser.is_group else mixin_bot_config.mixin_id
+    action = pack_input_action("source code", mixin_number)
+    pld = pack_button("Source code", action, BUTTON_COLORS[7])
+    return pld
+
+
+def get_source_code():
+    return "Source code: https://github.com/infowoods/oogway-mixin-bot"
+
+
 def get_help_doc():
     text = "## Help\n"
     text += "> Oogway, it is a querier. Be used by Mixin MUI(Message User Interface) as yet.\n"
@@ -34,4 +45,7 @@ def get_help_doc():
     text += "* `asset <symbol>`, search asset by symbol. e.g. `asset btc`\n"
     text += "* `wiki <title>`, search wikipedia\n"
     text += "* `size <item> <model>`, query size of Paper, iPhone, etc.\n"
+    text += "* `tr`/`翻`/`translate`/`翻译`, translation\n"
+
+    text += "\n\nSource code: https://github.com/infowoods/oogway-mixin-bot"
     return text