+++ message_handler/__init__.py
@@ -1,1 +1,1 @@
-from .index import handle, error_callback
+from .index import handle