+++ message_handler/from_user.py
@@ -1,52 +1,40 @@
-from mixinsdk.types.message import (
-    MessageView,
-    pack_post_data,
-    pack_text_data,
-    pack_message,
-)
+import logging
 
-
-from thisbot.init import (
-    logger,
-    mixin_bot_config,
-    mixin_client,
-    operation,
-)
-from thisbot.types import MessageUser
-from thisbot.constants import APP_NAME
 import basic_reply
-
 import commander
+from mixinsdk.types.message import (MessageView, pack_message, pack_post_data,
+                                    pack_text_data)
+from thisbot.constants import APP_NAME
+from thisbot.init import mixin_bot_config, mixin_client, operation
+from thisbot.types import MessageUser
 
 
-async def send_text_to_user(text, conversation_id, quote_message_id=None):
-    await mixin_client.blaze.send_message(
+def send_text_to_user(text, conversation_id, quote_message_id=None):
+    mixin_client.blaze.send_message(
         pack_message(
             pack_text_data(text), conversation_id, quote_message_id=quote_message_id
         )
     )
 
 
-async def send_post_to_user(text, conversation_id, quote_message_id=None):
-    await mixin_client.blaze.send_message(
+def send_post_to_user(text, conversation_id, quote_message_id=None):
+    mixin_client.blaze.send_message(
         pack_message(
             pack_post_data(text), conversation_id, quote_message_id=quote_message_id
         )
     )
 
 
-async def notice_operator(text):
-    logger.warn(f"notice operator: {text}")
+def notice_operator(text):
+    logging.warn(f"notice operator: {text}")
     text = f"{APP_NAME}:\n```" + text + "\n```"
-    await mixin_client.blaze.send_message(
-        pack_post_data(text), operation.notice_conv_id
-    )
+    mixin_client.blaze.send_message(pack_post_data(text), operation.notice_conv_id)
 
 
-async def handle_text(msguser: MessageUser, msgview: MessageView):
+def handle_text(msguser: MessageUser, msgview: MessageView):
     # limit size
     if len(msgview.data_decoded) > 8192:
-        await send_text_to_user("✗ Command too long", msgview.conversation_id)
+        send_text_to_user("✗ Command too long", msgview.conversation_id)
         return
 
     msg_text = msgview.data_decoded
@@ -57,21 +45,21 @@ async def handle_text(msguser: MessageUser, msgview: MessageView):
     sign_str = "@" + mixin_bot_config.mixin_id + " "
     cmd = cmd.replace(sign_str, "")
 
-    logger.info(f"user command: {cmd}")
+    logging.info(f"user command: {cmd}")
 
     ctx = commander.CommandContext(msguser, msgview)
-    await commander.handle(ctx, cmd)
+    commander.handle(ctx, cmd)
 
     for msg in ctx.replying_msgs:
-        await mixin_client.blaze.send_message(msg)
+        mixin_client.blaze.send_message(msg)
 
 
-async def handle_media(msguser: MessageUser, msgview: MessageView):
+def handle_media(msguser: MessageUser, msgview: MessageView):
     text = f"{msgview.category}\n\n```\n{msgview.data_decoded}\n```"
-    await send_post_to_user(text, msgview.conversation_id)
+    send_post_to_user(text, msgview.conversation_id)
 
 
-async def handle_conversation_actions(msgview: MessageView):
+def handle_conversation_actions(msgview: MessageView):
     # print(msgview.data_decoded)
     participant_id = msgview.data_decoded.get("participant_id")
     if participant_id != mixin_client.blaze.config.client_id:
@@ -79,16 +67,16 @@ async def handle_conversation_actions(msgview: MessageView):
     # 是对本机器人账号的操作才关心和处理
     action = msgview.data_decoded.get("action")
     if action == "ADD":
-        logger.info(f"被加入了群聊会话: {msgview.conversation_id}")
+        logging.info(f"被加入了群聊会话: {msgview.conversation_id}")
 
         text = basic_reply.get_welcome()
-        await send_text_to_user(text, msgview.conversation_id)
+        send_text_to_user(text, msgview.conversation_id)
 
         return
     elif action == "REMOVE":
-        logger.info(f"被加入了群聊会话: {msgview.conversation_id}")
+        logging.info(f"被加入了群聊会话: {msgview.conversation_id}")
         # log to db?
         return
     else:
-        logger.warn(f"未知的会话操作: {action}")
+        logging.warn(f"未知的会话操作: {action}")
         return