+++ message_handler/index.py
@@ -1,18 +1,15 @@
+import logging
+
+from mixinsdk.clients.blaze_client import BlazeClient
 from mixinsdk.types.message import MESSAGE_CATEGORIES, MessageView
-from thisbot.init import logger, mixin_client
-from thisbot.types import MessageUser
 from thisbot.constants import USER_TYPES
+from thisbot.types import MessageUser
 
 from . import from_operator, from_user
-from .methods import get_ext_user_type_and_id, get_mixin_group_profile
+from .methods import get_mixin_group_profile, get_mixin_user_type_and_id
 
 
-def error_callback(error, details):
-    logger.error(error)
-    logger.error(details)
-
-
-async def handle(message):
+def handle(blaze: BlazeClient, message):
     action = message["action"]
 
     if action == "ACKNOWLEDGE_MESSAGE_RECEIPT":
@@ -24,8 +21,8 @@ async def handle(message):
         return
 
     if action == "ERROR":
-        logger.warn("--- received error message ---")
-        logger.warn(message)
+        logging.warn("--- received error message ---")
+        logging.warn(message)
         return
         """example message={
             "id": "00000000-0000-0000-0000-000000000000",
@@ -40,8 +37,8 @@ async def handle(message):
     if action == "CREATE_MESSAGE":
         error = message.get("error")
         if error:
-            logger.warn("--- received error message ---")
-            logger.warn(message)
+            logging.warn("--- received error message ---")
+            logging.warn(message)
             return
 
         data = message["data"]
@@ -54,32 +51,42 @@ async def handle(message):
         if msgview.type == "message":
             # === from system message
             if msgview.user_id == "00000000-0000-0000-0000-000000000000":
-                logger.debug("received system message")
-                logger.debug(msgview.to_dict())
+                logging.debug("received system message")
+                logging.debug(msgview.to_dict())
                 if msgview.category == MESSAGE_CATEGORIES.SYSTEM_CONVERSATION:
-                    await from_user.handle_conversation_actions(msgview)
-                    await mixin_client.blaze.echo(msgview.message_id)
+                    from_user.handle_conversation_actions(msgview)
+                    blaze.echo(msgview.message_id)
                     return
 
-                logger.warn(f"unknown system message category: {msgview.category}")
+                logging.warn(f"unknown system message category: {msgview.category}")
                 return
 
             msguser = MessageUser()
             (
                 msguser.ext_user_type,
                 msguser.ext_user_id,
-            ) = get_ext_user_type_and_id(msgview.conversation_id, msgview.user_id)
-            logger.debug(
-                f"{msgview.category} message from: {msguser.ext_user_type} - {msguser.ext_user_id}"
-            )
+            ) = get_mixin_user_type_and_id(msgview.conversation_id, msgview.user_id)
+            msg = f"Message {msgview.category} - {msgview.message_id} From: {msguser.ext_user_type} - {msguser.ext_user_id}"
+            logging.info(msg)
+            print(msg)
+
+            if not msguser.ext_user_type:
+                logging.warn(f"Ignore unsupported user's message: {msgview.to_dict()}")
+                blaze.echo(msgview.message_id)
+                return
+            if msguser.ext_user_type == USER_TYPES.MIXIN_APP:
+                # 过滤 APP USER 消息。不然机器人直接可能相互收发消息无止境
+                logging.info("Ignore mixin app user's message")
+                blaze.echo(msgview.message_id)
+                return
 
             if msgview.category == MESSAGE_CATEGORIES.SYSTEM_ACCOUNT_SNAPSHOT:
-                # logger.debug(msgview.data_decoded)
+                # logging.debug(msgview.data_decoded)
                 try:
-                    await from_user.handle_transfer(msguser, msgview)
-                    await mixin_client.blaze.echo(msgview.message_id)
+                    from_user.handle_transfer(msguser, msgview)
+                    blaze.echo(msgview.message_id)
                 except Exception as e:
-                    logger.exception("at snapshot message", exc_info=True)
+                    logging.exception("at snapshot message", exc_info=True)
                     raise e from None
                 return
 
@@ -91,23 +98,23 @@ async def handle(message):
                 get_mixin_group_profile(msguser, msgview.conversation_id)
                 if msguser.group_owner_id != msgview.user_id:
                     # filter out group messages not from owner
-                    logger.debug(
+                    logging.debug(
                         f"Ignore group message:{msgview.message_id} not from owner"
                     )
-                    await mixin_client.blaze.echo(msgview.message_id)
+                    blaze.echo(msgview.message_id)
                     return
 
             if msgview.category == MESSAGE_CATEGORIES.TEXT:
                 msg_text = msgview.data_decoded
                 if msg_text.startswith(":"):  # consider it is admin command
-                    await from_operator.handle_command(msguser, msgview)
+                    from_operator.handle_command(msguser, msgview)
                 else:
-                    await from_user.handle_text(msguser, msgview)
-                await mixin_client.blaze.echo(msgview.message_id)
+                    from_user.handle_text(msguser, msgview)
+                blaze.echo(msgview.message_id)
                 return
 
             if msgview.category == MESSAGE_CATEGORIES.POST:
-                await mixin_client.blaze.echo(msgview.message_id)
+                blaze.echo(msgview.message_id)
                 return
 
             if msgview.category in [
@@ -115,25 +122,25 @@ async def handle(message):
                 MESSAGE_CATEGORIES.AUDIO,
                 MESSAGE_CATEGORIES.VIDEO,
                 MESSAGE_CATEGORIES.FILE,
+                MESSAGE_CATEGORIES.STICKER,
             ]:
-                await from_user.handle_media(msguser, msgview)
-                await mixin_client.blaze.echo(msgview.message_id)
+                from_user.handle_media(msguser, msgview)
+                blaze.echo(msgview.message_id)
                 return
 
             if msgview.category in [
-                MESSAGE_CATEGORIES.STICKER,
                 MESSAGE_CATEGORIES.APP_CARD,
                 MESSAGE_CATEGORIES.LIVE,
             ]:
                 # ignore
-                logger.info(
+                logging.info(
                     f"Ignore message category: {msgview.category}, {msgview.message_id}"
                 )
-                await mixin_client.blaze.echo(msgview.message_id)
+                blaze.echo(msgview.message_id)
                 return
 
-            logger.warn(f"Unknown/Ignore message category: {message}")
-            await mixin_client.blaze.echo(msgview.message_id)
+            logging.warn(f"Unknown/Ignore message category: {message}")
+            blaze.echo(msgview.message_id)
             return
 
-    logger.warn("Unknown message:", message)
+    logging.warn("Unknown message:", message)