+++ mixinsdk/types/messenger_schema.py
@@ -87,6 +87,11 @@ def pack_sharing_live(height: int, width: int, url: str, thumb_url: str):
 
 
 def pack_input_action(text: str, at_mixin_number: str = None) -> str:
+    """
+    Arguments:
+    - at_mixin_number: str, optional,
+        use in group conversation to mention a specific user
+    """
     action = "input:"
     if at_mixin_number:
         action += f"@{at_mixin_number} "