+++ thisbot/constants.py
@@ -7,6 +7,7 @@ APP_NAME = "Oogway-Mixin-Bot"
 class _UserTypes:
     MIXIN_GROUP: str = "MIXIN_GROUP"
     MIXIN_USER: str = "MIXIN_USER"
+    MIXIN_APP: str = "MIXIN_APP"
 
 
 USER_TYPES = _UserTypes()