+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # CHANGELOG
 
-## 2022-05-01
+**2022-05-23**
+
+- Support for `translate` text (with Google Translate)
+
+**2022-05-01**
 
 - Support for query **Wikipedia** entries
 - Like CLI syntax, text command support for arguments and options
+
+**2022-04-29**
+
+- Support for query Mixin assets
+- Support for query Current (Mixin Messenger) conversation id and user id